requirements.txt
numpy
opencv-python
chess
stockfish
ikpy
pyFirmata
cv2-enumerate-cameras
chess

